requirements.txt
PyQt6==6.7.0
pandas==2.2.2
requests==2.32.3
SQLAlchemy==2.0.30
python-dotenv==1.0.1
pyodbc==5.1.0
openpyxl==3.1.2
xlrd==2.0.1
